ProductActivity: 상품 장바구니 담기 Room insert 로직 구현, 데이터베이스 스키마구조 일부 변경

+++ app/src/main/java/com/ssafy/cobaltcoffee/database/CartDao.kt
@@ -13,7 +13,7 @@ interface CartDao {
     suspend fun getCarts(userId: String): MutableList<CartDto>
 
     @Insert(onConflict = REPLACE)
-    suspend fun insertCart(dto: CartDto): Int
+    suspend fun insertCart(dto: CartDto): Long
 
     @Update
     suspend fun updateCart(dto: CartDto): Int
